requirements-minimal.txt
# ==============================================================================
# SAGE Framework — Minimal Requirements
# ==============================================================================
# For low-resource machines (CPU-only, 4 GB RAM, no GPU).
# Excludes: chromadb, sentence-transformers, langchain (heavy deps).
# RAG falls back to keyword search — fully functional, less semantic recall.
#
# Usage:
#   pip install -r requirements-minimal.txt
#   SAGE_MINIMAL=1 python src/main.py api --project medtech
# ==============================================================================

# Core
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# HTTP client
requests>=2.31.0
httpx>=0.26.0

# Data
pandas>=2.0.0
numpy>=1.26.0

# MCP server support
fastmcp>=0.1.0

# Hardware (optional — comment out if not needed)
pyserial>=3.5
# pylink-square>=0.12.0   # J-Link debugger

# Microsoft Teams auth (optional)
# msal>=1.26.0

# PowerPoint generation (optional)
# python-pptx>=0.6.21
